+++ packages/terra-graphs-docs/CHANGELOG.md
@@ -12,3 +12,4 @@
   * Added data objects and examples for Pie chart.
   * Added data objects and examples for Combination graph.
   * Added reflow examples for Line Graph.
+  * Added critical and basic example for timeline graph.